withdraw comments

+++ phase1/src/Withdrawable.java
@@ -1,4 +1,4 @@
 /** A withdrawable account can be withdraw amount of money*/
 public interface Withdrawable {
-      void withdraw(int amount);
+      void withdraw(double amount);
 }
